mvp-backend/requirements.txt
# Backend Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
